Added login with password
