+++ src/base-mixins/login.js
@@ -4,52 +4,18 @@
 *    - exposes the `userId` and `loggedIn` public properties
 */
 
-import * as multiStorage from "../lib/multi-storage.js";
-
-/*
-*   Private methods: they are invoked with the asteroid instance as context, but
-*   they are not exported so they don't clutter the Asteroid class prototype.
-*/
-
-function onLogin ({id, token}) {
-    this.userId = id;
-    this.loggedIn = true;
-    return multiStorage.set(this.endpoint + "__login_token__", token)
-        .then(this.emit.bind(this, "loggedIn"));
-}
-
-function onLogout () {
-    this.userId = null;
-    this.loggedIn = false;
-    return multiStorage.del(this.endpoint + "__login_token__")
-        .then(this.emit.bind(this, "loggedOut"));
-}
-
-function resumeLogin () {
-    return multiStorage.get(this.endpoint + "__login_token__")
-        .then(resume => {
-            if (!resume) {
-                throw new Error("No login token");
-            }
-            return {resume};
-        })
-        .then(this.login.bind(this))
-        .then(onLogin.bind(this))
-        .catch(onLogout.bind(this));
-}
+import {onLogin, onLogout, resumeLogin} from "../lib/login-method";
 
 /*
 *   Public methods
 */
 
 export function login (loginParameters) {
-    return this.call("login", loginParameters)
-        .then(onLogin.bind(this));
+    return this.call("login", loginParameters).then(onLogin.bind(this));
 }
 
 export function logout () {
-    return this.call("logout")
-        .then(onLogout.bind(this));
+    return this.call("logout").then(onLogout.bind(this));
 }
 
 /*
